% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_mutations.R
\name{add_mutation_types}
\alias{add_mutation_types}
\title{Add Mutation Type Column to Data Frame}
\usage{
add_mutation_types(
  df,
  mutation_col = "mutation",
  annotation_col = "annotation"
)
}
\arguments{
\item{df}{Data frame with a 'mutation' column}

\item{mutation_col}{Name of the column containing mutation strings}

\item{annotation_col}{Optional name of annotation column}
}
\value{
Data frame with added 'mutation_type' column
}
\description{
Adds a 'mutation_type' column to a data frame containing mutation data.
}
