% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{analyze_breseq}
\alias{analyze_breseq}
\title{Quick Analysis Wrapper}
\usage{
analyze_breseq(...)
}
\arguments{
\item{...}{Arguments passed to run_breseq_pipeline}
}
\value{
Analysis results
}
\description{
Simplified wrapper for common analysis scenarios.
}
