% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_position.R
\name{circular_distance}
\alias{circular_distance}
\title{Calculate Distance Between Positions on Circular Genome}
\usage{
circular_distance(pos1, pos2, genome_length)
}
\arguments{
\item{pos1}{First position}

\item{pos2}{Second position}

\item{genome_length}{Total genome length}
}
\value{
Shortest distance (always positive)
}
\description{
Calculates the shortest distance between two positions on a circular genome.
}
