% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_mutations.R
\name{classify_mutations}
\alias{classify_mutations}
\title{Classify Multiple Mutations (Vectorized)}
\usage{
classify_mutations(mutations, annotations = NULL)
}
\arguments{
\item{mutations}{Character vector of mutation strings}

\item{annotations}{Optional character vector of annotations}
}
\value{
Character vector of mutation types
}
\description{
Classify Multiple Mutations (Vectorized)
}
