% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutations.R
\name{create_circular_plot}
\alias{create_circular_plot}
\title{Create Circular Genome Plot}
\usage{
create_circular_plot(df, pos_col, color_by, title, genome_length)
}
\description{
Create Circular Genome Plot
}
\keyword{internal}
