% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutations.R
\name{create_linear_plot}
\alias{create_linear_plot}
\title{Create Linear Mutation Plot}
\usage{
create_linear_plot(df, pos_col, color_by, title, show_labels, genome_length)
}
\description{
Create Linear Mutation Plot
}
\keyword{internal}
