% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutations.R
\name{create_timeline_plot}
\alias{create_timeline_plot}
\title{Create Timeline Plot Showing Mutation Accumulation}
\usage{
create_timeline_plot(df, pos_col, color_by, title, stage_names)
}
\description{
Create Timeline Plot Showing Mutation Accumulation
}
\keyword{internal}
