% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{extract_single_html}
\alias{extract_single_html}
\title{Extract Tables from Single HTML File}
\usage{
extract_single_html(html_file)
}
\arguments{
\item{html_file}{Path to HTML file}
}
\value{
Named list of data frames
}
\description{
Internal function to extract tables from a single breseq HTML file.
}
\keyword{internal}
