% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_genbank.R
\name{format_number_with_commas}
\alias{format_number_with_commas}
\title{Format Number with Comma Separators}
\usage{
format_number_with_commas(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
Character string with commas (no extra spaces)
}
\description{
Format Number with Comma Separators
}
\keyword{internal}
