% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reference.R
\name{get_genome_lengths}
\alias{get_genome_lengths}
\title{Get Genome Lengths as Named Vector}
\usage{
get_genome_lengths(reference_file)
}
\arguments{
\item{reference_file}{Path to reference file}
}
\value{
Named numeric vector (names = seq_id, values = length)
}
\description{
Convenience function to get genome lengths in a format ready for use
with normalization functions.
}
\examples{
\dontrun{
genome_lengths <- get_genome_lengths("reference.gbk")
# chromosome      pBR322       pUC19
#    4600000        4361        2686
}
}
