% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_mutations.R, R/export_genbank.R
\name{get_mutation_color}
\alias{get_mutation_color}
\title{Get Mutation Type Color}
\usage{
get_mutation_color(mutation_type)

get_mutation_color(mutation_type)
}
\arguments{
\item{mutation_type}{Mutation type (INS, DEL, SNP, REP, UNKNOWN)}
}
\value{
Hex color code

Hex color code
}
\description{
Returns a color for each mutation type for visualization.

Returns hex color code for mutation type (same as circos visualization).
}
