% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reference.R
\name{get_replicon_topology}
\alias{get_replicon_topology}
\title{Get Replicon Topology as Named Vector}
\usage{
get_replicon_topology(reference_file)
}
\arguments{
\item{reference_file}{Path to reference file}
}
\value{
Named character vector (names = seq_id, values = topology)
}
\description{
Get Replicon Topology as Named Vector
}
