% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reference.R
\name{is_circular}
\alias{is_circular}
\title{Check if Replicon is Circular}
\usage{
is_circular(seq_id, reference_file)
}
\arguments{
\item{seq_id}{Sequence ID to check}

\item{reference_file}{Path to reference file}
}
\value{
Logical
}
\description{
Check if Replicon is Circular
}
