% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutations.R
\name{plot_mutation_summary}
\alias{plot_mutation_summary}
\title{Plot Mutation Type Summary}
\usage{
plot_mutation_summary(mutation_data, title = "Mutation Type Distribution")
}
\arguments{
\item{mutation_data}{Data frame with mutation data}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Creates a bar chart showing the count of each mutation type.
}
